seperated out routes from server file, added card styling
