+++ assets/css/styles.css
@@ -1,0 +1,8 @@
+body {
+    box-sizing: border-box;
+    margin: 0;
+}
+
+#test {
+    background-color: lightgray;
+}
