+++ views/index.handlebars
@@ -1,9 +1,9 @@
-<div class="container">
+<div class="container m-4">
     <div class="row justify-content-center">
-        <div class="col-md-8">
+        <div class="col-md-6">
             <h1>Welcome to Image Search</h1>
             <p>powered by Pexels API</p>
-            <a href="/favorites"><button class="btn btn-success">Saved Images</button></a>
+            <a href="/api/favorites"><button class="btn btn-success">Saved Images</button></a>
 
             <form action="/api" method="POST">
                 <div class="mb-3">
@@ -17,20 +17,26 @@
         </div>
     </div>
 
-    <div class="row mt-3">
-        <div class="col-md-12">
-            <div id="images">
-                {{#each allImages}}
-                <img src="{{ this.src.medium }}" alt="">
-                <p>Photograph By <a href="{{ this.photographer_url }}">{{ this.photographer }}</a></p>
-                <form action="/favorite" method="POST">
-                    <input type="hidden" name="photographer" value="{{ this.photographer }}">
-                    <input type="hidden" name="photographer_url" value="{{ this.photographer_url }}">
-                    <input type="hidden" name="photo_url" value="{{ this.src.medium }}">
-                    <button type="submit" class="btn btn-danger">Save</button>
-                </form>
-                {{/each }}
+    <div class="row m-4">
+        <div class="card-group">
+            {{#each allImages}}
+            <div class="col-sm-4">
+                <div class="card">
+                    <img src="{{ this.src.medium }}" class="card-img-top" alt="...">
+                    <div class="card-body">
+                        <p class="card-text">Photograph By <a href="{{ this.photographer_url }}">{{
+                                this.photographer}}</a>
+                        </p>
+                        <form action="/api/favorite" method="POST">
+                            <input type="hidden" name="photographer" value="{{ this.photographer }}">
+                            <input type="hidden" name="photographer_url" value="{{ this.photographer_url }}">
+                            <input type="hidden" name="photo_url" value="{{ this.src.medium }}">
+                            <button type="submit" class="btn btn-danger ">Save</button>
+                        </form>
+                    </div>
+                </div>
             </div>
+            {{/each }}
         </div>
     </div>
 </div>
